Add terminal login support
Add more nodes to /dev and passwd,group,inittab to /etc
add rcS, which mounts the contents of fstab
Make the contents of fstab mount on startup

This manually creates some device nodes that are needed for serial and
virtual terminals, as well as the user and group information so that it
is possible to log in as root.

It may be more appropriate to create a new morph, as this is not strictly
the fhs-dirs hierarchy, but at this stage that would require a new
repository and the alteration of foundation to require that morph.