config changes for [um]dev and ssh
Mdev is currently being used as it is part of busybox
This does not need all of the device nodes previously specified
as it is able to create them, so they are removed from create-dev-nodes
/dev/console and /dev/null may be needed to boot

rcS has been changed to use a tmpfs for /dev, this will be filled by
mdev. http://lwn.net/Articles/330985/ suggests that it may be faster
to use a devtmpfs as [um]dev only needs to set the permissions though.

More users and groups have been added. Many are for the standard
[um]dev device configuration. The sshd user and group are for ssh
as we do not have a good way to add groups in morph yet.