Add minimal /etc/services
The full list is between 188 and 750KB, which is too much to be in
foundation.

The current file is what is required for nfs to work, which would ideally
go in the nfs stratum, but since we can't merge files in /etc I have put
it here in fhs-dirs to avoid confusion when we add other network services.