Use 'devices' in morphology to create device nodes
Allowing arbitrary code to create device nodes is a bad idea, so
it is preferred to build under linux-user-chroot.

Since we need to create device nodes, a "devices" field is added to
the chunk morphology, which describes which devices nodes are to be
created, and how.

This commit will not work without the corresponding changes to morph.

Reviewed-by: Lars Wirzenius <lars.wirzenius@codethink.co.uk>